requirements.txt
addict==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
check-orientation==0.0.5
click==8.1.8
dlib==20.0.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
jpeg4py==0.1.4
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rich==14.3.2
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.24
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
